BIOSを実装
